refactor(store-core): single-source formatQuestionAnswerSummary (was duplicated app+dashboard)
Part of #6453. formatQuestionAnswerSummary + formatAnswerValue were duplicated in
the app (store/utils.ts) and dashboard (utils/questionAnswerSummary.ts), and had
DRIFTED — the dashboard version is a superset (handles the {otherLabel,
freeformText} freeform shape, #4651) while the app's is a deliberate subset
(#4755, mobile parity out of scope). Hoist the dashboard's canonical superset
into @chroxy/store-core/question-answer-summary and re-export from both clients.
Behaviour-preserving: the freeform branch is dormant on any client that never
produces that shape (the app passes only the Record shape), so the app's output
is unchanged. Test moved to store-core (17 pass); app (25) + dashboard (7) green;
all three packages typecheck clean.
